BIT-Website/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dateparser==0.7.2
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
python-dateutil==2.8.1
pytz==2019.3
regex==2020.1.8
six==1.14.0
soupsieve==1.9.5
tzlocal==2.0.0
urllib3==1.25.7
Werkzeug==0.16.0
